feat: add a setting flag to control user pref lang
This add a setting variable that allow to show if the LANGUAGE_CODE of the site config
would have more priority than the cookie of the user preference.
The setting is `IGNORE_USER_LANGUAGE_COOKIE`, so by default the site config `LANGUAGE_CODE`
woul have more prior but if the setting is set to False the user cookie would work.