feat: remove staff access for catalog views
This is controversial, but simple. To do this in a monkey patch is a bigger task and also with
more work. Is different due the change is in a nested function. Anyway the alternative is to
monkeyPatch the _dispatch method and handle there. But its risky and also that method manage other permission with different ways.