feat: remove staff access in courses for catalog views (#37)
* feat: remove staff access for catalog views

This is controversial, but simple. To do this in a monkey patch is a bigger task and also with
more work. Is different due the change is in a nested function. Anyway the alternative is to
monkeyPatch the _dispatch method and handle there. But its risky and also that method manage other permission with different ways.

* feat: keep custom test with default behavior

To use you configure the setting: `STAFF_CAN_SEE_IN_CATALOG`